Move Registration page to pages folder
